% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{check_metadata}
\alias{check_metadata}
\title{Check and arrange the data in the metadata}
\usage{
check_metadata(parameters, supported_file_types = c("rds", "qs2"))
}
\arguments{
\item{parameters}{a dataframe of the sample metadata info}

\item{supported_file_types}{file types to included in reports}
}
\value{
dataframe: a check passed metadata
}
\description{
Check and arrange the data in the metadata
}
\examples{
library(SeuratExplorerServer)
data_meta <- initialize_metadata(Reports.main = c(
system.file("extdata/source-data", "fly", package ="SeuratExplorerServer"),
system.file("extdata/source-data", "mouse", package ="SeuratExplorerServer")),
Rds.path = c("Rds-file/G101_PC20res04.rds", "haber.tsne.embeding.rds"),
Reports.second = c(NA, NA), Sample.name = c("Fly-Gut-EEs-scRNAseq-GuoXT",
"Mouse-Intestine-scRNAseq-Haber"))
invisible(check_metadata(parameters = data_meta))
}
