% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{initialize_metadata}
\alias{initialize_metadata}
\title{initialize sample metadata}
\usage{
initialize_metadata(Reports.main, Rds.path, Reports.second, Sample.name)
}
\arguments{
\item{Reports.main}{primary analysis directory, Rds should be included, and all files located in this directory will be included in reports}

\item{Rds.path}{relative path of rds/qs2 file under the primary analysis directory}

\item{Reports.second}{secondary analysis directory, and all files located in this directory will also be included in reports, such 'cellranger' outputs will be linked to Others directory}

\item{Sample.name}{Sample name}
}
\value{
A data.frame
}
\description{
When building a new app, it is used to initialise the metadata, which records the primary and secondary analysis
directories for the data, the path of the rds/qs2 file under the primary analysis directory and the name of each sample;
the primary analysis directory should contain the Rds/qs2 file.
}
\examples{
data_meta <- initialize_metadata(Reports.main = c(
system.file("extdata/demo", "fly", package ="SeuratExplorerServer"),
system.file("extdata/demo", "mouse", package ="SeuratExplorerServer")),
Rds.path = c("Rds-file/G101_PC20res04.rds", "haber.tsne.embeding.rds"),
Reports.second = c(NA, NA), Sample.name = c("Fly-Gut-EEs-scRNAseq-GuoXT",
"Mouse-Intestine-scRNAseq-Haber"))
data_meta
# saveRDS(data_meta,file = system.file("extdata/demo/others",
# "sample-paramters.rds", package ="SeuratExplorerServer"))

}
