% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server.R
\name{server}
\alias{server}
\title{Shiny Server}
\usage{
server(input, output, session)
}
\arguments{
\item{input}{Input from the UI}

\item{output}{Output to send back to UI}

\item{session}{shiny session}
}
\value{
shiny server functions
}
\description{
Shiny Server
}
