% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui.R
\name{ui}
\alias{ui}
\title{UI for shiny App interface}
\usage{
ui(
  Encrypted.app,
  TechnicianEmail = "zhangyongchao@nibs.ac.cn",
  TechnicianName = "Zhang Yongchao"
)
}
\arguments{
\item{Encrypted.app}{whether to encrypt app}

\item{TechnicianEmail}{Email of the technician}

\item{TechnicianName}{Name of the technician}
}
\value{
shiny UI
}
\description{
UI for shiny App interface
}
