% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/launch.R
\name{launchSeuratExplorerServer}
\alias{launchSeuratExplorerServer}
\title{Launch shiny app}
\usage{
launchSeuratExplorerServer(
  Encrypted = TRUE,
  credentials = data.frame(user = "shiny", password = "12345", stringsAsFactors = FALSE),
  parameterfile = revise_demo_path(),
  TechnicianEmail = "zhangyongchao@nibs.ac.cn",
  TechnicianName = "ZhangYongchao",
  ReductionKeyWords = c("umap", "tsne"),
  ReportsFileTypes = c("pdf", "tiff", "tif", "jpeg", "gif", "jpg", "png", "bmp", "svg",
    "html", "mp4", "avi", "Rmd", "R", "ipynb", "sh", "Sh", "txt", "csv", "xlsx", "xls",
    "xml", "md", "py"),
  DefaultSplitMaxLevel = 6,
  SupportedFileTypes = c("rds", "qs2"),
  CommentsFile = NULL,
  TechnicianUser = "admin",
  verbose = FALSE
)
}
\arguments{
\item{Encrypted}{whether to encrypt app}

\item{credentials}{a data frame with the credentials}

\item{parameterfile}{path to the parameter file(rds), a data frame with columns:Reports.main, Rds.path, Reports.second, Sample.name
SplitOptions.MaxLevel, Default.DimensionReduction, Default.ClusterResolution, Species, and Description}

\item{TechnicianEmail}{Email of the technician}

\item{TechnicianName}{Name of the technician}

\item{ReductionKeyWords}{keywords to extract reductions for the reduction options}

\item{ReportsFileTypes}{File types to be included in reports}

\item{DefaultSplitMaxLevel}{the max factor level of the column from metadata to be included in split option}

\item{SupportedFileTypes}{supported file types}

\item{CommentsFile}{path to the CSV file that stores the comment board.
Defaults to a \code{comments.csv} file next to \code{parameterfile}.}

\item{TechnicianUser}{character vector of login usernames with technician
privileges (allowed to delete comments / mark them resolved). Ignored when
\code{Encrypted = FALSE} (everyone is treated as technician).}

\item{verbose}{default FALSE, messages for debug use}
}
\value{
In-browser Shiny Application launch
}
\description{
Launch shiny app
}
\examples{
if(interactive()){launchSeuratExplorerServer()}
}
