% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/launch.R
\name{launchSeuratExplorerServer}
\alias{launchSeuratExplorerServer}
\title{Launch shiny app}
\usage{
launchSeuratExplorerServer(
  Encrypted = TRUE,
  credentials = data.frame(user = "shiny", password = "12345", stringsAsFactors = FALSE),
  paramterfile = system.file("extdata/demo/others", "sample-paramters.rds", package =
    "SeuratExplorerServer"),
  TechnicianEmail = "zhangyongchao@nibs.ac.cn",
  TechnicianName = "ZhangYongchao"
)
}
\value{
In-browser Shiny Application launch
}
\description{
Launch shiny app
}
\examples{
# launchSeuratExplorerServer()
}
