% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/launch.R
\name{launchSeuratExplorerServer}
\alias{launchSeuratExplorerServer}
\title{Launch shiny app}
\usage{
launchSeuratExplorerServer(
  Encrypted = TRUE,
  credentials = data.frame(user = "shiny", password = "12345", stringsAsFactors = FALSE),
  paramterfile = revise_demo_path(),
  TechnicianEmail = "zhangyongchao@nibs.ac.cn",
  TechnicianName = "ZhangYongchao",
  ReportsFileTypes = c("pdf", "tiff", "tif", "jped", "jpg", "png", "bmp", "svg"),
  DefaultSplitMaxLevel = 6,
  SupportedFileTypes = c("rds", "qs2"),
  verbose = FALSE
)
}
\arguments{
\item{Encrypted}{weather to encrypt app}

\item{credentials}{a data frame with the credentials}

\item{paramterfile}{path to the parameter file(rds), a data frame with columns:Reports.main, Rds.path, Reports.second, Sample.name
SplitOptions.MaxLevel, Default.DimensionReduction, Default.ClusterResolution, Species, and Description}

\item{TechnicianEmail}{Email of the technician}

\item{TechnicianName}{Name of the technician}

\item{ReportsFileTypes}{File types to be included in reports}

\item{DefaultSplitMaxLevel}{the max factor level of the column from metadata to be included in split option}

\item{SupportedFileTypes}{supported file types}

\item{verbose}{TRUE for debug use}
}
\value{
In-browser Shiny Application launch
}
\description{
Launch shiny app
}
\examples{
if(interactive()){launchSeuratExplorerServer()}
}
