% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{revise_demo_path}
\alias{revise_demo_path}
\title{Revise demo data path}
\usage{
revise_demo_path(
  paramterfile = system.file("extdata", "data_meta.rds", package =
    "SeuratExplorerServer")
)
}
\arguments{
\item{paramterfile}{the path to metadata file}
}
\value{
path to parameter file
}
\description{
Revise demo data path
}
\examples{
revise_demo_path()
}
