% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server.R
\name{server}
\alias{server}
\title{Server for SeuratExplorer shiny app}
\usage{
server(input, output, session)
}
\arguments{
\item{input}{Input from the UI}

\item{output}{Output to send back to UI}

\item{session}{from shiny server function}
}
\description{
Server for SeuratExplorer shiny app
}
