% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui.R
\name{ui}
\alias{ui}
\title{UI for shiny App interface}
\usage{
ui(
  Encrypted.app,
  TechnicianEmail = "zhangyongchao@nibs.ac.cn",
  TechnicianName = "ZhangYongchao"
)
}
\arguments{
\item{Encrypted.app}{是否要加密App}

\item{TechnicianEmail}{技术人员的联系邮箱}

\item{TechnicianName}{技术人员的名字}
}
\description{
UI for shiny App interface
}
